ML models and their performance
I have created a classification dataset with sklearn (make_classification) . I have tested different classifiers to check which is performing better and consumes less time.
Light GBM has performed magnificently with high accuracy and speed. Still it depends on the nature of the problem, the dataset with which you are working. If speed is the criteria then go for Microsoft LightGBM and if accuracy is the criteria then offcourse XGBoost is ahead in the race